// src/sdk/AuroraEventList.ts
export enum AuroraEventList {
    usbConnectionChange = "usbConnectionChange",
    bluetoothConnectionChange = "bluetoothConnectionChange",
    findBluetoothDevice = "findBluetoothDevice",
    cmdBegin = "cmdBegin",
    cmdEnd = "cmdEnd",
    msdAttachmentChange = "msdAttachmentChange",
    flashConnectionChange = "flashConnectionChange",
    cmdInputRequested = "cmdInputRequested",
    cmdOutputReady = "cmdOutputReady",
    log = "log",
    streamData = "streamData",
    auroraEvent = "auroraEvent",
    auroraError = "auroraError",
}